refactor: Hide implementation details about global database.
