+++ pyproject.toml
@@ -8,7 +8,7 @@ classifiers = ["License :: OSI Approved :: MIT License"]
 keywords = ["snowflake", "snowflakedb", "fake", "local", "mock", "testing"]
 requires-python = ">=3.9"
 dependencies = [
-    "duckdb~=0.9.2",
+    "duckdb~=0.10.0",
     "pyarrow",
     "snowflake-connector-python",
     "sqlglot~=21.1.0",
