+++ tests/test_users.py
@@ -1,7 +1,5 @@
 import snowflake.connector.cursor
 
-from fakesnow.connection import USERS_TABLE_FQ_NAME
-
 
 def test_show_users_base_case(cur: snowflake.connector.cursor.SnowflakeCursor):
     result = cur.execute("SHOW USERS")
@@ -10,7 +8,8 @@ def test_show_users_base_case(cur: snowflake.connector.cursor.SnowflakeCursor):
 
 
 def test_show_users_with_users(cur: snowflake.connector.cursor.SnowflakeCursor):
-    cur.execute(f"INSERT INTO {USERS_TABLE_FQ_NAME} (name) VALUES ('foo'), ('bar')")
+    result = cur.execute("CREATE USER foo")
+    result = cur.execute("CREATE USER bar")
 
     result = cur.execute("SHOW USERS")
     assert result
